fix: widen score fields from u32 to u64 to prevent silent overflow
Discovered while trying to find a genuine "top record" for the AI: a
20-game headless benchmark (dellacherie evaluator, 200k-piece cap)
reached a mean score of ~4.01 billion, only ~270M below u32::MAX
(4,294,967,295). Score's u32 field, combined with an uncapped level
multiplier (level = lines_cleared/10 + 1, no ceiling) and this
workspace not enabling overflow-checks in release builds, meant any
long enough game — AI or human — would silently wrap the score to a
small, wrong number with no error or panic. No wraparound had actually
happened yet in that run (all 20 scores were tightly clustered and
consistent with piece/line counts, no anomalous outliers), but pushing
the piece cap any higher to find a real completed game would have hit
it blind.

Fix: GameState.score, Well.score, CrossGame::total_score(), and
scoring::line_clear_score() are now u64 (score itself can't realistically
reach u64::MAX in any game). The wasm bridge exposes score as f64 rather
than u32/u64 — JS `number` represents integers exactly up to 2^53, far
beyond any reachable score, avoiding wasm-bindgen's u64->bigint mapping
which would have forced every web/ call site to switch from `number` to
`bigint`. sim's GameResult.score and percentile() widened to match.

New regression test (engine/tests/line_clear_tests.rs): sets score to
u32::MAX - 50, applies a score-earning action, and asserts the result
is u32::MAX - 50 + 100 (not a wrapped small number) and > u32::MAX.

cargo test --workspace: 86 passed (16 suites, up from 84 — the new
overflow test plus one line_clear_tests count fix).

+++ ai/src/dellacherie.rs
@@ -210,7 +210,7 @@ pub fn play_best_cross_move(cross: &mut CrossGame, weights: &Weights) {
     let drop_distance = (placement.row - active.row).max(0) as u32;
     cross.force_active_placement(placement.rotation, placement.row, placement.column);
     cross.apply(Action::HardDrop);
-    cross.wells[placement.arm.index()].score += drop_distance * engine::scoring::HARD_DROP_POINTS_PER_CELL;
+    cross.wells[placement.arm.index()].score += drop_distance as u64 * engine::scoring::HARD_DROP_POINTS_PER_CELL;
 }
 
 /// Single-board equivalent of `best_cross_placement`, mirroring
